WW-5616 - JakartaStreamMultiPartRequest warns on file delete if the file doesnt exist (#1591)
* Pull aspects into alignment with main/7.x+ AbstractMultiPartRequest.java
* Update JakartaMultiPartRequest and JakartaStreamMultiPartRequest to use isFile()
* Update cleanup text to mirror main/7.x+

+++ core/src/main/java/org/apache/struts2/dispatcher/multipart/JakartaStreamMultiPartRequest.java
@@ -79,10 +79,16 @@ public class JakartaStreamMultiPartRequest extends AbstractMultiPartRequest {
                 try {
                     // catch any exceptions during cleanup to ensure all files are deleted.
                     File file = fileInfo.getFile();
-                    LOG.debug("Deleting file '{}'.", file.getName());
-                    if (!file.delete()) {
-                        LOG.warn("There was a problem attempting to delete file [{}].",
-                            file.getName());
+                    if(file != null) {
+                        if(file.isFile()) {
+                            LOG.debug("Deleting file: {}", file.getName());
+                            if (!file.delete()) {
+                                LOG.warn("There was a problem attempting to delete file: {}",
+                                    file.getName());
+                            }
+                        } else {
+                            LOG.debug("File: {} already deleted", file.getName());
+                        }
                     }
                 } catch (Exception e) {
                     LOG.warn("Error during cleanup of file item: [{}]", normalizeSpace(fileInfo.getOriginalName()), e);
